Fix analyze view for mavlink 2.0.

+++ src/ui/MAVLinkDecoder.cc
@@ -68,14 +68,13 @@ void MAVLinkDecoder::run()
 void MAVLinkDecoder::receiveMessage(LinkInterface* link,mavlink_message_t message)
 {
     if (message.msgid >= cMessageIds) {
-        // No support for messag ids above 255
         return;
     }
 
     Q_UNUSED(link);
     memcpy(receivedMessages+message.msgid, &message, sizeof(mavlink_message_t));
 
-    uint8_t msgid = message.msgid;
+    uint32_t msgid = message.msgid;
     const mavlink_message_info_t* msgInfo = mavlink_get_message_info(&message);
 
     // Store an arrival time for this message. This value ends up being calculated later.
@@ -220,11 +219,11 @@ void MAVLinkDecoder::emitFieldValue(mavlink_message_t* msg, int fieldid, quint64
     if (componentMulti[msg->msgid] == true) multiComponentSourceDetected = true;
 
     // Add field tree widget item
-    uint8_t msgid = msg->msgid;
+    uint32_t msgid = msg->msgid;
     if (messageFilter.contains(msgid)) return;
     QString fieldName(msgInfo->fields[fieldid].name);
     QString fieldType;
-    uint8_t* m = (uint8_t*)&((mavlink_message_t*)(receivedMessages+msgid))->payload64[0];
+    uint32_t* m = (uint32_t*)&((mavlink_message_t*)(receivedMessages+msgid))->payload64[0];
     QString name("%1.%2");
     QString unit("");
 
@@ -324,7 +323,7 @@ void MAVLinkDecoder::emitFieldValue(mavlink_message_t* msg, int fieldid, quint64
     case MAVLINK_TYPE_UINT8_T:
         if (msgInfo->fields[fieldid].array_length > 0)
         {
-            uint8_t* nums = m+msgInfo->fields[fieldid].wire_offset;
+            uint32_t* nums = m+msgInfo->fields[fieldid].wire_offset;
             fieldType = QString("uint8_t[%1]").arg(msgInfo->fields[fieldid].array_length);
             for (unsigned int j = 0; j < msgInfo->fields[fieldid].array_length; ++j)
             {
